refactor: cookie policy page location
